@@ -11,19 +11,18 @@
   <packaging>jar</packaging>
 
   <properties>
-    <jdk.version>1.8</jdk.version>
+    <jdk.version>17</jdk.version>
 
     <dependency.version.commons-codec>1.8</dependency.version.commons-codec>
     <dependency.version.commons-lang>2.6</dependency.version.commons-lang>
     <dependency.version.commons-logging>1.1.3</dependency.version.commons-logging>
     <dependency.version.fop>1.1</dependency.version.fop>
-    <dependency.version.guava>18.0</dependency.version.guava>
-    <dependency.version.guice>4.0</dependency.version.guice>
+    <dependency.version.guava>30.1-jre</dependency.version.guava>
+    <dependency.version.guice>5.1.0</dependency.version.guice>
     <dependency.version.jcommander>1.30</dependency.version.jcommander>
     <dependency.version.jetty>9.2.6.v20141205</dependency.version.jetty>
     <dependency.version.jsr305>2.0.1</dependency.version.jsr305>
     <dependency.version.log4j>2.1</dependency.version.log4j>
-    <dependency.version.moxy>2.6.0</dependency.version.moxy>
     <dependency.version.saxon>9.6.0-4</dependency.version.saxon>
     <dependency.version.servlet-api>3.1.0</dependency.version.servlet-api>
     <dependency.version.sitebricks>0.8.10</dependency.version.sitebricks>
@@ -180,11 +179,6 @@
       </exclusions>
     </dependency>
     <dependency>
-      <groupId>com.google.inject.extensions</groupId>
-      <artifactId>guice-multibindings</artifactId>
-      <version>${dependency.version.guice}</version>
-    </dependency>
-    <dependency>
       <groupId>com.google.code.findbugs</groupId>
       <artifactId>jsr305</artifactId>
       <version>${dependency.version.jsr305}</version>
@@ -259,10 +253,14 @@
     </dependency>
 
     <dependency>
+      <groupId>jakarta.xml.bind</groupId>
+      <artifactId>jakarta.xml.bind-api</artifactId>
+      <version>3.0.1</version>
+    </dependency>
+    <dependency>
       <groupId>org.eclipse.persistence</groupId>
       <artifactId>org.eclipse.persistence.moxy</artifactId>
-      <version>${dependency.version.moxy}</version>
-      <scope>compile</scope>
+      <version>4.0.6</version>
     </dependency>
 
     <dependency>
