@@ -1,6 +1,6 @@
 package me.romankh.resumegenerator.model;
 
-import javax.xml.bind.annotation.*;
+import jakarta.xml.bind.annotation.*;
 import java.util.ArrayList;
 import java.util.List;
 
