@@ -1,6 +1,6 @@
 package me.romankh.resumegenerator.model;
 
-import javax.xml.bind.annotation.XmlAttribute;
+import jakarta.xml.bind.annotation.XmlAttribute;
 
 /**
  * @author Roman Khmelichek
