@@ -6,5 +6,5 @@
 @XmlAccessorType(XmlAccessType.NONE)
 package me.romankh.resumegenerator.model;
 
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
+import jakarta.xml.bind.annotation.XmlAccessType;
+import jakarta.xml.bind.annotation.XmlAccessorType;
