@@ -4,7 +4,7 @@ import me.romankh.resumegenerator.model.Resume;
 import me.romankh.resumegenerator.parser.ResumeParser;
 import org.xml.sax.SAXException;
 
-import javax.xml.bind.JAXBException;
+import jakarta.xml.bind.JAXBException;
 import java.io.IOException;
 import java.io.InputStream;
 
